add error for productlink, add data in data.sql

+++ src/main/java/com/openclassroom/projet12/service/CategorieService.java
@@ -1,11 +1,15 @@
 package com.openclassroom.projet12.service;
 
 import com.openclassroom.projet12.dto.CategorieDTO;
+import com.openclassroom.projet12.exceptions.ErrorCode;
+import com.openclassroom.projet12.exceptions.NotEmptyException;
 import com.openclassroom.projet12.exceptions.NotFoundException;
 import com.openclassroom.projet12.mapper.CategorieMapper;
 import com.openclassroom.projet12.model.Categorie;
+import com.openclassroom.projet12.model.Product;
 import com.openclassroom.projet12.model.SubCategorie;
 import com.openclassroom.projet12.respository.CategorieRepository;
+import com.openclassroom.projet12.respository.ProductRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,8 +25,10 @@ import static java.util.stream.Collectors.toList;
 @AllArgsConstructor
 public class CategorieService {
 
+
     private final CategorieRepository categorieRepository;
     private final SubCategorieService subCategorieService;
+    private final ProductRepository productRepository;
 
     public List<CategorieDTO> getCategories() {
         return categorieRepository.findAll()
@@ -40,7 +46,7 @@ public class CategorieService {
 
     public Categorie getCategorie(Long id) {
         return categorieRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("La catégorie n'existe pas !"));
+                .orElseThrow(() -> new NotFoundException("La catégorie n'existe pas !", ErrorCode.CATEGORIE_NOT_FOUND_ERROR));
     }
 
     public Categorie addCategorie(CategorieDTO categorieDTO) {
@@ -56,6 +62,11 @@ public class CategorieService {
     }
 
     public Long deleteCategorie(Long id) {
+        Categorie categorie = getCategorie(id);
+        List<Product> list = productRepository.findAllByCategorie(categorie);
+        if(!list.isEmpty()) {
+            throw new NotEmptyException("La catégorie est toujours liée à un produit", ErrorCode.PRODUCT_LINKED_ERROR);
+        }
         categorieRepository.deleteById(id);
         return id;
     }
