+++ src/main/java/com/openclassroom/projet12/service/SubCategorieService.java
@@ -1,9 +1,13 @@
 package com.openclassroom.projet12.service;
 
 import com.openclassroom.projet12.dto.SubCategorieDTO;
+import com.openclassroom.projet12.exceptions.ErrorCode;
+import com.openclassroom.projet12.exceptions.NotEmptyException;
 import com.openclassroom.projet12.exceptions.NotFoundException;
 import com.openclassroom.projet12.mapper.SubCategorieMapper;
+import com.openclassroom.projet12.model.Product;
 import com.openclassroom.projet12.model.SubCategorie;
+import com.openclassroom.projet12.respository.ProductRepository;
 import com.openclassroom.projet12.respository.SubCatagorieRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,10 +22,11 @@ import java.util.stream.Collectors;
 public class SubCategorieService {
 
     private final SubCatagorieRepository subCatagorieRepository;
+    private final ProductRepository productRepository;
 
     public SubCategorie getSubCategorie(Long id) {
         return subCatagorieRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("La sous-catégorie du product n'éxiste pas"));
+                .orElseThrow(() -> new NotFoundException(("La sous-catégorie du product n'existe pas"), ErrorCode.SUBCATEGORIE_NOT_FOUND_ERROR));
     }
 
     public List<SubCategorieDTO> getSubCategories() {
@@ -51,6 +56,11 @@ public class SubCategorieService {
     }
 
     public Long deleteSubCategorie(Long id) {
+        SubCategorie subCategorie = getSubCategorie(id);
+        List<Product> list = productRepository.findAllBySubCategorie(subCategorie);
+        if(!list.isEmpty()) {
+            throw new NotEmptyException("La sous-catégorie est toujours lié à un produit", ErrorCode.PRODUCT_LINKED_ERROR);
+        }
         subCatagorieRepository.deleteById(id);
         return id;
     }
