+++ src/main/java/com/openclassroom/projet12/service/TagService.java
@@ -1,9 +1,13 @@
 package com.openclassroom.projet12.service;
 
 import com.openclassroom.projet12.dto.TagDTO;
+import com.openclassroom.projet12.exceptions.ErrorCode;
+import com.openclassroom.projet12.exceptions.NotEmptyException;
 import com.openclassroom.projet12.exceptions.NotFoundException;
 import com.openclassroom.projet12.mapper.TagMapper;
+import com.openclassroom.projet12.model.Product;
 import com.openclassroom.projet12.model.Tag;
+import com.openclassroom.projet12.respository.ProductRepository;
 import com.openclassroom.projet12.respository.TagRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -17,6 +21,7 @@ import java.util.List;
 public class TagService {
 
     private final TagRepository tagRepository;
+    private final ProductRepository productRepository;
 
 
     public List<Tag> getTags() {
@@ -33,7 +38,7 @@ public class TagService {
     }
     public Tag getTag(Long id) {
         return tagRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Le tag recherché n'a pas été trouvé"));
+                .orElseThrow(() -> new NotFoundException(("Le tag recherché n'a pas été trouvé"), ErrorCode.TAG_NOT_FOUND_ERROR));
     }
 
     public Tag addTag(TagDTO tagDTO) {
@@ -47,7 +52,15 @@ public class TagService {
     }
 
     public Long deleteTag(Long id) {
+        List<Product> list = productRepository.getProductsByTagId(id);
+        if(!list.isEmpty()) {
+            throw new NotEmptyException("Le tag est toujours lié à un produit", ErrorCode.PRODUCT_LINKED_ERROR);
+        }
         tagRepository.deleteById(id);
         return id;
+    }
+
+    public Tag findByName(String name) {
+       return tagRepository.findTagByName(name);
     }
 }
